Fixed some broken unit tests

+++ src/Libraries/SmartStore.Data/Migrations/201508042207146_RemoveKeepAlive.cs
@@ -1,14 +1,19 @@
 namespace SmartStore.Data.Migrations
 {
-    using System;
-    using System.Data.Entity.Migrations;
+	using System;
+	using System.Data.Entity.Migrations;
+	using System.Web.Hosting;
+	using Core.Data;
 	using SmartStore.Data.Setup;
 
 	public partial class RemoveKeepAlive : DbMigration, ILocaleResourcesProvider, IDataSeeder<SmartObjectContext>
     {
         public override void Up()
         {
-			Sql("DELETE FROM [dbo].[ScheduleTask] WHERE [Type] = 'SmartStore.Services.Common.KeepAliveTask, SmartStore.Services'");
+			if (HostingEnvironment.IsHosted)
+			{
+				Sql("DELETE FROM [dbo].[ScheduleTask] WHERE [Type] = 'SmartStore.Services.Common.KeepAliveTask, SmartStore.Services'");
+			}
         }
         
         public override void Down()
